fixed ampIO
